% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praiseR.R
\name{praise}
\alias{praise}
\title{Deliver praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{This is the name of the person you want to praise. Input as text string.}

\item{punctuation}{This is the punctuation at the end of the praise. Default is "!".}
}
\value{
Text string with personalised praise
}
\description{
A function that praises the intended name.
}
\examples{
praise(name = "Erika", punctuation = "!")
}
